# api/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "milmil API",
    "version": "1.0.0",
    "description": "Self-hosted anime media server"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/user/preferences/backup-config": {
      "get": {
        "summary": "List backup configurations",
        "operationId": "listBackupConfigs",
        "tags": [
          "User Preferences"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of backup configs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Upsert a backup configuration",
        "operationId": "upsertBackupConfig",
        "tags": [
          "User Preferences"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Config saved"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API token (mlml_ prefix)",
        "description": "Pass an API token obtained from POST /api/v1/auth/login or POST /api/v1/api-tokens. Tokens have the prefix `mlml_`."
      }
    },
...
  },
...
}